refactor(hook): rename useUser to useLoggedInUser
