+++ components/Footer.js
@@ -11,7 +11,7 @@ import styled from 'styled-components';
 
 import languages from '../lib/constants/locales';
 
-import { useUser } from '../components/UserProvider';
+import { useLoggedInUser } from '../components/UserProvider';
 
 import TranslateIcon from './icons/TranslateIcon';
 import Container from './Container';
@@ -221,7 +221,7 @@ const Footer = () => {
   const intl = useIntl();
   const languageOptions = React.useMemo(generateLanguageOptions);
   const defaultLanguage = languageOptions.find(language => language.value === intl.locale);
-  const { LoggedInUser } = useUser();
+  const { LoggedInUser } = useLoggedInUser();
   const formatLanguageOptionLabel = ({ value, label }, { context }) => (
     <Span fontSize="12px" fontWeight={context === 'menu' && value === intl.locale ? 'bold' : 'normal'}>
       {label}
