+++ components/admin-panel/sections/AccountSettings.js
@@ -15,10 +15,10 @@ import { adminPanelQuery } from '../../../pages/admin-panel';
 import SettingsForm from '../../edit-collective/Form';
 import Loading from '../../Loading';
 import { TOAST_TYPE, useToasts } from '../../ToastProvider';
-import { useUser } from '../../UserProvider';
+import { useLoggedInUser } from '../../UserProvider';
 
 const AccountSettings = ({ account, section }) => {
-  const { LoggedInUser, refetchLoggedInUser } = useUser();
+  const { LoggedInUser, refetchLoggedInUser } = useLoggedInUser();
   const router = useRouter();
   const [state, setState] = React.useState({ status: undefined, result: undefined });
   const { addToast } = useToasts();
