+++ components/admin-panel/sections/AuthorizedApps.js
@@ -15,14 +15,14 @@ import { authorizedAppsQuery } from '../../oauth/queries';
 import Pagination from '../../Pagination';
 import StyledHr from '../../StyledHr';
 import { P } from '../../Text';
-import { useUser } from '../../UserProvider';
+import { useLoggedInUser } from '../../UserProvider';
 
 const AuthorizedAppsSection = () => {
   const router = useRouter() || {};
   const query = router.query;
   const variables = { limit: 10, offset: query.offset ? parseInt(query.offset) : 0 };
   const { data, loading, error, refetch } = useQuery(authorizedAppsQuery, { variables, context: API_V2_CONTEXT });
-  const { LoggedInUser } = useUser();
+  const { LoggedInUser } = useLoggedInUser();
   const authorizations = data?.loggedInAccount?.oAuthAuthorizations;
 
   // Redirect to previous page when removing the last item of a page
