+++ components/host-dashboard/AddFundsModal.js
@@ -34,7 +34,7 @@ import StyledSelect from '../StyledSelect';
 import StyledTooltip from '../StyledTooltip';
 import { P, Span } from '../Text';
 import { TOAST_TYPE, useToasts } from '../ToastProvider';
-import { useUser } from '../UserProvider';
+import { useLoggedInUser } from '../UserProvider';
 
 import illustration from '../contribution-flow/fees-on-top-illustration.png';
 
@@ -248,7 +248,7 @@ const getTiersOptions = (intl, tiers) => {
 };
 
 const AddFundsModal = ({ host, collective, ...props }) => {
-  const { LoggedInUser } = useUser();
+  const { LoggedInUser } = useLoggedInUser();
   const [fundDetails, setFundDetails] = useState({});
   const { addToast } = useToasts();
   const intl = useIntl();
