+++ components/orders/OrdersWithData.js
@@ -20,7 +20,7 @@ import Pagination from '../Pagination';
 import SearchBar from '../SearchBar';
 import StyledHr from '../StyledHr';
 import { H1 } from '../Text';
-import { useUser } from '../UserProvider';
+import { useLoggedInUser } from '../UserProvider';
 
 import OrdersFilters from './OrdersFilters';
 import OrdersList from './OrdersList';
@@ -152,7 +152,7 @@ const OrdersWithData = ({ accountSlug, title, status, showPlatformTip }) => {
   const queryVariables = { accountSlug, ...getVariablesFromQuery(router.query, status) };
   const queryParams = { variables: queryVariables, context: API_V2_CONTEXT };
   const { data, error, loading, variables, refetch } = useQuery(accountOrdersQuery, queryParams);
-  const { LoggedInUser } = useUser();
+  const { LoggedInUser } = useLoggedInUser();
   const prevLoggedInUser = usePrevious(LoggedInUser);
 
   // Refetch data when user logs in
