+++ lib/hooks/withLoggedInUser.js
@@ -4,10 +4,10 @@ import * as Sentry from '@sentry/browser';
 import dayjs from 'dayjs';
 import jwt from 'jsonwebtoken';
 
-import { loggedInUserQuery } from './graphql/queries';
-import { refreshToken, refreshTokenWithTwoFactorCode, refreshTokenWithTwoFactorRecoveryCode } from './api';
-import { getFromLocalStorage, LOCAL_STORAGE_KEYS, setLocalStorage } from './local-storage';
-import LoggedInUser from './LoggedInUser';
+import { refreshToken, refreshTokenWithTwoFactorCode, refreshTokenWithTwoFactorRecoveryCode } from '../api';
+import { loggedInUserQuery } from '../graphql/queries';
+import { getFromLocalStorage, LOCAL_STORAGE_KEYS, setLocalStorage } from '../local-storage';
+import LoggedInUser from '../LoggedInUser';
 
 const maybeRefreshAccessToken = async currentToken => {
   const decodeResult = jwt.decode(currentToken);
