+++ pages/guest-join.js
@@ -20,7 +20,7 @@ import StyledButton from '../components/StyledButton';
 import StyledCard from '../components/StyledCard';
 import StyledRadioList from '../components/StyledRadioList';
 import { H4, P, Span } from '../components/Text';
-import { useUser } from '../components/UserProvider';
+import { useLoggedInUser } from '../components/UserProvider';
 
 const STATUS = {
   SUBMITTING: 'SUBMITTING',
@@ -192,7 +192,7 @@ const JoinAsGuest = () => {
 
 const JoinGuestPage = () => {
   const intl = useIntl();
-  const { LoggedInUser } = useUser();
+  const { LoggedInUser } = useLoggedInUser();
 
   return (
     <Page title={intl.formatMessage(MESSAGES.pageTitle, { service: 'Open Collective' })}>
