+++ pages/welcome.js
@@ -9,7 +9,7 @@ import Image from '../components/Image';
 import Link from '../components/Link';
 import StyledCard from '../components/StyledCard';
 import StyledLink from '../components/StyledLink';
-import { useUser } from '../components/UserProvider';
+import { useLoggedInUser } from '../components/UserProvider';
 
 const WelcomeOptionContainer = styled(Container)`
   &:hover {
@@ -18,7 +18,7 @@ const WelcomeOptionContainer = styled(Container)`
 `;
 
 const Welcome = () => {
-  const { LoggedInUser } = useUser();
+  const { LoggedInUser } = useLoggedInUser();
   const intl = useIntl();
 
   return (
